My first project
On this code, I was trying to put in practice some ideas that I had a couple days ago, but isn't complete. I need to learn how to add results that are already inside a "var".
This is a very simple code. I'm just testing some skills that I have on this start.